Moving MPINormalCalculationUtils to mpi core.
