+++ kratos/mpi/utilities/mpi_normal_calculation_utilities.h
